Added dependencies as individual refs
